backend/requirements.txt
boto3==1.26.69
botocore==1.29.69
jmespath==1.0.1
numpy==1.24.2
Pillow==9.4.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
